PC_IO.ASM: fix 64KiB bounary DMA write support
The commit 6945c7102e56d1b92ef29848d6a56cc154f5053a was incorrect.  It
did not handle write access properly.  In case of read, issuing
command is first.  In case of write, copying data is first.

In addition, accessing order is changed for performance.  For example,
if accessing order is sector 1-4 -> 5 -> 6-9, the drive might need to
rotate the disk 3 times.  If accessing order is sector 1-4 -> 6-9 ->
5, the sector 6 might be found soon.  The idea is so, but not tested
on real drives.