IO.ASM: fix OUT issues
Previous OUT routine had several issues.  I found the following
issues, compared to the v2.0/bin/MSDOS.SYS binary in the MS-DOS source
tree:

- AH=2 did not check ^C.
- Output with AH=40H to cooked mode CON did not check ^C.
- Tab position was incorrect if control character other than 9, 10, 13
  was printed.
- ^P did not print anything to a printer.

I found an OUT routine in v1.25 source and used it as a reference to
find and fix the issue.  Variable CHARCO which I found in MSDATA.ASM
was not used anywhere, so I guessed IO.ASM should use it.
Unfortunately the v1.25 source does not have the variable, the code
behaves like the MSDOS.SYS binary about the STATCHK call (from IO.SYS
view), but the variable usage may be different.