PC_IO, PC_BOOT: PCjr support
360KiB disk format used by PCjr is now supported.

On Japanese version of JX, bit 6 of BIOS drive number which is set by
default on boot means accessing 80 cylinders (720KiB) disk.  When
accessing 360KiB (40 cylinders) PCjr disk, the bit needs to be
cleared.  Previously the bit was always set on JX.  Now media ID value
is used for setting the bit properly.  Therefore accessing 360KiB disk
on JX is now supported, and booting from 360KiB disk on JX is also
expected to be supported.